+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ignored HMAC character capitalization
   ([#93](https://github.com/laterpay/laterpay-client-python/issues/93))
+* Added support for ``/dialog/subscribe`` and LaterPay's subscription system
 
 ## 5.0.0